[dependencies] Upgrade to LLVM 11.0.0 (#200)
* [dependencies] Upgrade to LLVM 11.0.0

* [api] Map AttributeIndex to Int instead of Long

The presets for LLVM 11 return an Int for the AttributeIndex signatures now which means the AttributeIndex enum should be updated.

+++ src/main/kotlin/io/vexelabs/bitbuilder/llvm/ir/Flags.kt
@@ -141,13 +141,13 @@ public enum class AtomicRMWBinaryOperation(public override val value: Int) :
     }
 }
 
-public enum class AttributeIndex(public override val value: Long) :
-    ForeignEnum<Long> {
+public enum class AttributeIndex(public override val value: Int) :
+    ForeignEnum<Int> {
     Return(LLVM.LLVMAttributeReturnIndex),
     Function(LLVM.LLVMAttributeFunctionIndex);
 
-    public companion object : ForeignEnum.CompanionBase<Long, AttributeIndex> {
-        public override val map: Map<Long, AttributeIndex> by lazy {
+    public companion object : ForeignEnum.CompanionBase<Int, AttributeIndex> {
+        public override val map: Map<Int, AttributeIndex> by lazy {
             values().associateBy(AttributeIndex::value)
         }
     }
